requirements.txt
Flask > 0.8
Pillow
pytest
jsonschema
azure-cognitiveservices-vision-face